requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-extensions==3.2.3
prisma==0.12.0
python-dotenv==1.0.0
dj-database-url==2.2.0
cryptography==41.0.7
django-oauth-toolkit==1.7.1
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7
drf-spectacular==0.26.5
faker==20.1.0
django-model-utils==4.3.1


